Genericize files, add num_paths recursive solution+tests
